Fix a warning..
git-svn-id: https://piggledy.org/svn/peerfuse/trunk@733 f7b84f85-2e04-4de5-9e99-c4612dc0539c

+++ common/cache_base.cpp
@@ -15,7 +15,7 @@ int CacheInterface::Read(std::string path, char* buf, size_t size, off_t off)
 {
 	FileContent& file = content_list.GetFile(path);
 
-	if(off > file.GetFileSize())
+	if(off > (off_t)file.GetFileSize())
 		return 0;
 
 	/* Limit the read to the size of the file */
